road_project_manager/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
django-environ==0.11.2
Pillow>=10.0.1
boto3>=1.29.7
django-storages>=1.14.2
gunicorn>=21.2.0
whitenoise>=6.6.0
djangorestframework-gis>=1.0
django-filter>=23.3